Move Question Cell implementations into their own files
