+++ pretixSCAN/Ticket Scanning/QuestionCells/CountryCodeQuestionCell.swift
@@ -1,0 +1,14 @@
+//
+//  CountryCodeQuestionCell.swift
+//  pretixSCAN
+//
+//  Created by Daniel Jilg on 13.08.19.
+//  Copyright © 2019 rami.io. All rights reserved.
+//
+
+import UIKit
+
+class CountryCodeQuestionCell: QuestionCell {
+    override class var reuseIdentifier: String { return "CountryCodeQuestionCell" }
+    // TODO
+}
